fix: use u16 for listen port end-to-end
Port numbers are 16-bit; the codebase had listen_port/listener_port
typed as u32, inconsistent with PeerAddress.port (already u16).
Migrate the type all the way through the CLI, config, ClientSettings,
Client, ServerActor, and the listener entry point. The wire format is
preserved: build_set_wait_port_message now takes u16 and widens to u32
via .into() before write_int32. Also rename DEFALT_LISTEN_PORT to
DEFAULT_LISTEN_PORT.

+++ soulseek-rs-lib/src/actor/server_actor.rs
@@ -129,7 +129,7 @@ pub enum ServerMessage {
 pub struct ServerActor {
     address: PeerAddress,
     context: Arc<RwLock<Context>>,
-    listen_port: u32,
+    listen_port: u16,
     enable_listen: bool,
     stream: Option<TcpStream>,
     connection_state: ConnectionState,
@@ -146,7 +146,7 @@ impl ServerActor {
     pub fn new(
         address: PeerAddress,
         client_channel: Sender<ClientOperation>,
-        listen_port: u32,
+        listen_port: u16,
         enable_listen: bool,
     ) -> Self {
         Self {
